requirements.txt
asgiref==3.4.0
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
django-paypal==1.1.1
django-phonenumber-field==5.2.0
idna==2.10
phonenumbers==8.12.27
Pillow==8.3.1
pip==21.1.3
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
setuptools==52.0.0.post20210125
six==1.16.0
sqlparse==0.4.1
stripe==2.58.0
typing-extensions==3.10.0.0
urllib3==1.26.6
wheel==0.36.2
